+++ elasticsearch/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
 
-VERSION = (0, 4, 3)
+VERSION = (0, 4, 4)
 __version__ = VERSION
 __versionstr__ = '.'.join(map(str, VERSION))